+++ lib/feature/employee/presentation/pages/update_employee.dart
@@ -7,6 +7,7 @@ import 'package:pluto_grid/pluto_grid.dart';
 
 import '../../../../core/constants/app_colors.dart';
 import '../../../../core/functions/hijri_picker.dart';
+import '../../../../core/utils/helper_method.dart';
 import '../../../../core/widgets/custom_button.dart';
 import '../../../../core/widgets/custom_check_box.dart';
 import '../../../../core/widgets/custom_dropdown_button.dart';
@@ -870,14 +871,21 @@ class UpdateEmployee extends StatelessWidget {
                     }).onTap(controller.pickImage).paddingAll(15),
                     CustomButton(
                       text: 'تعديل',
-                      onPressed: () => controller.save(),
+                      onPressed: () {
+                        if (checkUpdatePermission()) {
+                          controller.save();
+                        }
+                      },
                       height: 30,
                       width: 80,
                     ),
                     CustomButton(
                       text: 'حذف',
-                      onPressed: () =>
-                          controller.delete(int.parse(controller.id.text)),
+                      onPressed: () {
+                        if (checkDeletePermission()) {
+                          controller.delete(int.parse(controller.id.text));
+                        }
+                      },
                       height: 30,
                       width: 80,
                     ),
